requirements.txt
# requirements.txt  (for Dockerfile.base)
apache-airflow-providers-postgres
psycopg2-binary
pandas
sqlalchemy
requests